Create Player and Agent class
Made the project more modular by creating a Player and Agent class. All scanner interaction with the player is now handled by the Player class. All of the logic of the Agent will be handled in the Agent class. The BattleshipGame class now merely handles the inputs of Player and Agent.